refactor: unify LLM JSON parsing into shared parse_llm_json function
Move _parse_llm_json and _repair_truncated_json from agent.py to
llm_client.py as parse_llm_json (public).  All five ad-hoc
strip+json.loads patterns in radar.py and patrol.py now use the
shared parser, gaining truncated JSON repair support.
