+++ src/repokeeper/agent.py
@@ -14,7 +14,6 @@ from __future__ import annotations
 
 import json
 import os
-import re
 from pathlib import Path
 from typing import Any
 
@@ -26,7 +25,7 @@ from repokeeper.git_ops import (
     apply_and_push,  # noqa: F401  # re-export
 )
 from repokeeper.git_ops import git as _git  # noqa: F401  # re-export
-from repokeeper.llm_client import LLMClient, TokenUsage
+from repokeeper.llm_client import LLMClient, TokenUsage, parse_llm_json
 from repokeeper.logs import get_logger
 from repokeeper.profile import load_profile
 from repokeeper.repo_context import (  # noqa: F401  # re-export
@@ -199,118 +198,6 @@ Respond with a single valid JSON object — no markdown fences, no explanation o
 """
 
 
-def _parse_llm_json(raw: str) -> dict[str, Any]:
-    """Parse LLM JSON output with resilience to common formatting issues.
-
-    Handles markdown fences, code block markers, and attempts basic repair
-    for unterminated strings.
-
-    Args:
-        raw: Raw text content from LLM response.
-
-    Returns:
-        Parsed JSON dict.
-
-    Raises:
-        ValueError: If the response could not be parsed as JSON.
-    """
-    text = raw.strip()
-
-    # ── Strip markdown code fences ──
-    fence_pattern = r"```(?:json)?\s*\n(.*?)\n```"
-    m = re.search(fence_pattern, text, re.DOTALL)
-    if m:
-        text = m.group(1).strip()
-    elif text.startswith("```"):
-        inner = text.split("```", 2)
-        if len(inner) >= 2:
-            candidate = inner[1]
-            if candidate.startswith("json"):
-                candidate = candidate[4:]
-            text = candidate.strip()
-
-    # ── Try direct parse ──
-    try:
-        return json.loads(text)  # type: ignore[no-any-return]
-    except json.JSONDecodeError as err:
-        first_error = err
-
-    # ── Try extracting the outermost JSON object ──
-    start = text.find("{")
-    end = text.rfind("}")
-    if start != -1 and end != -1 and end > start:
-        extracted = text[start : end + 1]
-        try:
-            return json.loads(extracted)  # type: ignore[no-any-return]
-        except json.JSONDecodeError:
-            pass
-
-    # ── Try fixing common unterminated string ──
-    repaired = _repair_truncated_json(text)
-    if repaired is not None:
-        try:
-            return json.loads(repaired)  # type: ignore[no-any-return]
-        except json.JSONDecodeError:
-            pass
-
-    raise ValueError(
-        f"Failed to parse LLM JSON response. First error: {first_error}\n"
-        f"Raw response (last 2000 chars): ...{text[-2000:]}"
-    )
-
-
-def _repair_truncated_json(text: str) -> str | None:
-    """Attempt to repair a truncated/incomplete JSON string.
-
-    Tries to find the last complete key-value pair and close the object.
-
-    Args:
-        text: Possibly truncated JSON string.
-
-    Returns:
-        Repaired string or ``None`` if repair is not possible.
-    """
-    bracket_stack: list[str] = []
-    in_string = False
-    escape_next = False
-
-    for ch in text:
-        if escape_next:
-            escape_next = False
-            continue
-        if ch == "\\":
-            escape_next = True
-            continue
-        if ch == '"' and not escape_next:
-            in_string = not in_string
-            continue
-        if in_string:
-            continue
-        if ch in "{[":
-            bracket_stack.append(ch)
-        elif ch == "}":
-            if bracket_stack and bracket_stack[-1] == "{":
-                bracket_stack.pop()
-        elif ch == "]":
-            if bracket_stack and bracket_stack[-1] == "[":
-                bracket_stack.pop()
-
-    if not bracket_stack and not in_string:
-        return None
-
-    repaired = text.rstrip()
-    if in_string:
-        repaired += '"'
-
-    for open_ch in reversed(bracket_stack):
-        if open_ch == "{":
-            repaired += "}"
-        elif open_ch == "[":
-            repaired += "]"
-
-    return repaired if repaired != text else None
-
-
 def call_llm(
     issue_data: dict[str, Any],
     context_str: str,
@@ -394,7 +281,7 @@ def call_llm(
         raw = response.content.strip()
 
         try:
-            return _parse_llm_json(raw), total_usage
+            return parse_llm_json(raw), total_usage
         except ValueError as err:
             if attempt < max_retries:
                 logger.warning(
